add security group for RDS and update env vars
The infrastructure was missing a security group allowing the Airflow
instance to communicate with the data warehouse, I added a security
group allowing TCP traffic from the Airflow instance security group.

The name of the environment variables containing values for the finance
scraper (especially variables relevant for the data warehouse) were
outdated and needed to be updated in the 'config.py' module.