Update dashboard and define mappings for web page - company
I fix dashboard data processing on the "worst 10" table, and update the
SQL command fetching the data to remove null values.

Work is in progress for mapping company names to web pages. This is
necessary because the web pages have been update to remove the company
name from most pages.