collection date = Airflow date + 1 business day (#43)
Airflow's DAG run date is the date at the start of the schedule
interval, so this is one day before the actual date at which the job
runs for a daily DAG. We store the 'collection date' in the database,
which was set with the DAG run date, so it was wrong. Now we add one
business day to the DAG run date to obtain the collection date.